99_その他/03_ツール/KindleOCR/requirements.txt
pyautogui
pygetwindow
pytesseract
Pillow
opencv-python
